DigRec/requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
tensorflow>=2.8.0
PyQt5>=5.15.0
pillow>=8.0.0
